% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_col_names.R
\name{change_col_names}
\alias{change_col_names}
\title{Create custom column names for pivotting.}
\usage{
change_col_names(x)
}
\arguments{
\item{x}{not used, but need by pivot_wider()}
}
\value{
A character vector with custom column names.
}
\description{
Create nice column names when pivoting MRM data.
}
\author{
Rico derks
}
