% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sim_data.R
\name{extract_sim_data}
\alias{extract_sim_data}
\title{Extract the SIM traces}
\usage{
extract_sim_data(data, sim_names, sim_ids)
}
\arguments{
\item{data}{MChromatograms object from readSRMData() containing all the sims from your experiment}

\item{sim_names}{A character vector containing the names you want to give to each trace.}

\item{sim_ids}{a integer vector containing the index of each sim traces you want.}
}
\value{
A data frame containing the retention time and intensity of each sim trace in long format.
}
\description{
Extract the SIM traces from all the data and put them in a nice data frame.
}
\author{
Rico Derks
}
