% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find all peaks}
\usage{
find_peaks(
  data,
  sim_ids,
  sim_names,
  noise,
  peakwidth = c(0.15, 0.5),
  snthresh = 10
)
}
\arguments{
\item{data}{MChromatograms object from readSRMData() containing all the sims from your experiment}

\item{sim_ids}{a integer vector containing the index of each sim traces you want.}

\item{sim_names}{A character vector containing the names you want to give to each trace.}

\item{noise}{a numeric vector containg the noise levels for each trace.}

\item{peakwidth}{numeric(2) with the expected approximate peak width in chromatographic space. Given as a range (min, max) in minutes.}

\item{snthresh}{numeric(1) defining the signal to noise ratio cutoff.}
}
\value{
A data frame with all the peaks found in each trace in long format.
}
\description{
Perfrom peak picking (CentWave) on all traces.
}
\author{
Rico Derks
}
