% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_mzml.R
\name{fix_mzml}
\alias{fix_mzml}
\title{Fix mzML file}
\usage{
fix_mzml(mzml_file, new_mzml_file)
}
\arguments{
\item{mzml_file}{character vector containing the filename of the mzML.}

\item{new_mzml_file}{character vector containing the filename of the new to
create mzML file.}
}
\description{
Fix a mzML file with SIM data for reading.
}
\details{
mzML files made with more recent versions of ProteoWizard can generate
    the following error: `Error: Can not open file <filename>! Original error
    was: Error in pwizModule$open(filename): [IO::HandlerBinaryDataArray]
    Unknown binary data type.`

    This is error is caused by the usage of an older ProteoWizard library by the
    package mzR (for more information see https://github.com/lgatto/MSnbase/issues/517).

    This function uses the `xml2` package to remove the part which is giving
    problems.
}
\author{
Rico Derks
}
