% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_sim.R
\name{get_all_sim}
\alias{get_all_sim}
\title{Show all the SIM traces.}
\usage{
get_all_sim(data)
}
\arguments{
\item{data}{MChromatograms object from readSRMData().}
}
\value{
A 1 column matrix with the *m/z* values of all SIM traces.
}
\description{
Get all the *m/z* values of all SIM traces from a MChromatograms object.
}
\author{
Rico Derks
}
