% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_peaks.R
\name{group_peaks}
\alias{group_peaks}
\title{Group peaks}
\usage{
group_peaks(peak_data, rt_diff = 5)
}
\arguments{
\item{peak_data}{a data frame with all the peaks found in each trace in long format.}

\item{rt_diff}{the retention time window to group around in seconds. Default is 5 seconds}
}
\value{
A data frame the same as the input, but with some extra columns added containing the grouping information.
}
\description{
Group the peaks over all the traces
}
\author{
Rico Derks
}
