% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cim.R
\name{plot_cim}
\alias{plot_cim}
\title{titla Plot CIM chromatogram}
\usage{
plot_cim(
  sim_data,
  peak_list,
  select_sim = NULL,
  rt_line = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{sim_data}{Data frame containing all the SIM traces (from extract_sims()).}

\item{peak_list}{Data frame containing peak information (from group_peaks), cleaned up.}

\item{select_sim}{A character vector, which SIM trace to show.}

\item{rt_line}{A numerical value. If not NULL a dashed line is shown at this position.}

\item{title}{A character vector. The title of the plot.}

\item{subtitle}{A character vector. The subtitle of the plot.}
}
\value{
A ggplot graph with the full CIM chromatogram and a zoom part
}
\description{
Create a correlation ion monitoring chromatogram.
}
\author{
Rico Derks
}
