% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reconstruct_spectrum.R
\name{plot_reconstruct_spectrum}
\alias{plot_reconstruct_spectrum}
\title{Plot the reconstructed EISA fragmentation spectrum}
\usage{
plot_reconstruct_spectrum(
  peak_list,
  peak_group = NULL,
  title = NULL,
  subtitle = NULL,
  xmin = 20,
  xmax = 260,
  xstep = 20
)
}
\arguments{
\item{peak_list}{data frame with the peak list}

\item{peak_group}{integer, for which peak group you want to plot the reconstructed
EISA fragmentation spectrum}

\item{title}{tilte of the plot}

\item{subtitle}{subtitle of the plot}

\item{xmin}{start of the *m/z* axis}

\item{xmax}{end of the *m/z* axis}

\item{xstep}{step size of the *m/z* axis}
}
\value{
A ggplot object containing the reconstructed EISA fragmentation spectrum.
}
\description{
Plot the reconstructed EISA fragmentation spectrum.
}
\author{
Rico Derks
}
