% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_spectrum.R
\name{reconstruct_spectrum}
\alias{reconstruct_spectrum}
\title{Calculate the ratio's for the reconstructed EISA fragmentation spectrum}
\usage{
reconstruct_spectrum(peak_list)
}
\arguments{
\item{peak_list}{data frame with the peak list}
}
\value{
The same data frame, but with additional columns `intens_perc` and `intens_ratio`.
}
\description{
The ratio's of the precursor and the 3 fragments are calculate so a
    reconstructed EISA fragmentation spectrum can be created.
}
\author{
Rico Derks
}
